+++ src/main/kotlin/blackjack/domain/deck/Deck.kt
@@ -1,4 +1,4 @@
-package blackjack.domain
+package blackjack.domain.deck
 
 import blackjack.domain.card.Card
 import java.util.LinkedList
