+++ src/main/kotlin/blackjack/domain/gameResult/PlayerResult.kt
@@ -1,5 +1,5 @@
 package blackjack.domain.gameResult
 
-import blackjack.domain.Player
+import blackjack.domain.participant.Player
 
 data class PlayerResult(val player: Player, val status: GameResultStatus)
