+++ src/test/kotlin/blackjack/ParticipantTest.kt
@@ -1,9 +1,9 @@
 package blackjack
 
-import blackjack.domain.Participant
 import blackjack.domain.card.Card
 import blackjack.domain.card.Rank
 import blackjack.domain.card.Suit
+import blackjack.domain.participant.Participant
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
