+++ src/test/kotlin/blackjack/PlayerTest.kt
@@ -1,9 +1,9 @@
 package blackjack
 
-import blackjack.domain.Player
 import blackjack.domain.card.Card
 import blackjack.domain.card.Rank
 import blackjack.domain.card.Suit
+import blackjack.domain.participant.Player
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
